backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.3.7
pydantic==2.5.0
python-jose==3.3.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.2 